GLOB-5138: Add a URIField
Includes:
 -  Reorganization of custom fields
 -  Addition of the `rfc3986` library and a URIField
 -  Elimination of the `urltools` library that makes an HTTP request to the Internet on load
 -  Movement of URIMatcher to an importable location
