+++ microcosm_flask/tests/test_namespaces.py
@@ -11,9 +11,9 @@ from hamcrest import (
 from mock import Mock
 
 from microcosm.api import create_object_graph
+from microcosm_flask.matchers import matches_uri
 from microcosm_flask.namespaces import Namespace
 from microcosm_flask.operations import Operation
-from microcosm_flask.tests.matchers import matches_url
 
 
 def test_endpoint_for():
@@ -136,7 +136,7 @@ def test_operation_href_for_qs():
 
     with graph.app.test_request_context():
         url = ns.href_for(Operation.Search, offset=0, limit=10, qs=dict(foo="bar"))
-        assert_that(url, matches_url("http://localhost/api/foo?offset=0&limit=10&foo=bar"))
+        assert_that(url, matches_uri("http://localhost/api/foo?offset=0&limit=10&foo=bar"))
 
 
 def test_namespace_accepts_controller():
